requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-dotenv>=1.0.1
pydantic>=2.7.0

langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langgraph>=0.2.30

chromadb>=0.5.5
tavily-python>=0.5.0
sentence-transformers>=3.0.0

httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.0

pypdf>=5.0.0
python-docx>=1.1.0

pytest>=8.3.0
